requirements.txt
fastapi==0.115.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
crewai==0.28.5
langchain>=0.1.10,<0.2.0
langchain-core>=0.1.28,<0.2.0
langchain-google-genai==0.0.6




